Cleaned some code of the UI, and fixed all the warnings
