+++ app/src/main/java/com/example/characterDevelopment/ui/CharactersList/CharactersListScreen.kt
@@ -16,7 +16,6 @@
 
 package com.example.characterDevelopment.ui.CharactersList
 
-import android.util.Log
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -45,19 +44,18 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.example.compose.rally.ui.components.AppTabsRow
-import com.example.compose.rally.ui.components.CharacterRow
-import com.example.compose.rally.ui.components.ConfirmationDialog
-import com.example.compose.rally.ui.components.dropDownMenuIcon
+import com.example.characterDevelopment.ui.components.AppTabsRow
+import com.example.characterDevelopment.ui.components.CharacterRow
+import com.example.characterDevelopment.ui.components.ConfirmationDialog
+import com.example.characterDevelopment.ui.components.dropDownMenuIcon
 import com.example.characterDevelopment.R
-import com.example.characterDevelopment.data.database.entities.Order
 import com.example.characterDevelopment.data.database.entities.orders
 import com.example.characterDevelopment.domain.Models.CharacterDomainModel
 import com.example.characterDevelopment.ui.CharacterLevel.CharacterScreen
 import com.example.characterDevelopment.ui.ViewModels.CharacterCreatorViewModel
 import com.example.characterDevelopment.ui.ViewModels.SettingsViewModel
 import com.example.characterDevelopment.ui.Views.AppSettings
-import com.example.characterDevelopment.ui.Views.AppDestinations
+import com.example.characterDevelopment.ui.Views.AppScreens
 import com.example.characterDevelopment.ui.Views.CharactersList
 import com.example.characterDevelopment.ui.Views.MainCharacter
 import com.example.characterDevelopment.ui.Views.addCharacterRoute
@@ -66,14 +64,14 @@ import com.example.characterDevelopment.ui.Views.updateTextsLanguage
 import com.example.characterDevelopment.ui.appConfiguration.ConfigurationScreen
 
 @Composable
-fun characterListScreen(
+fun CharacterListScreen(
     navController: NavController,
     charactersVm: CharacterCreatorViewModel,
     settingsVm: SettingsViewModel
 ) {
     updateTextsLanguage()
 
-    var currentScreen: AppDestinations by remember { mutableStateOf(CharactersList) }
+    var currentScreen: AppScreens by remember { mutableStateOf(CharactersList) }
 
     //After changing configuration, sometimes the app gets restarted, in this case return to config screen
     if (settingsVm.updateConfiguration) {
@@ -85,7 +83,7 @@ fun characterListScreen(
         topBar = {
             AppTabsRow(
                 allScreens = appTabRowScreen,
-                onTabSelected = { screen: AppDestinations ->
+                onTabSelected = { screen: AppScreens ->
                     currentScreen = screen
                 },
                 currentScreen = currentScreen,
