+++ app/src/main/java/com/example/characterDevelopment/ui/appConfiguration/ConfigurationScreen.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.example.compose.rally.ui.components.dropDownMenuRowWithConfirmation
+import com.example.characterDevelopment.ui.components.dropDownMenuRowWithConfirmation
 import com.example.characterDevelopment.R
 import com.example.characterDevelopment.data.database.entities.AppLanguage
 import com.example.characterDevelopment.data.database.entities.Theme
@@ -24,7 +24,7 @@ import com.example.characterDevelopment.domain.Models.SettingsDomainModel
 import com.example.characterDevelopment.ui.ViewModels.CharacterCreatorViewModel
 import com.example.characterDevelopment.ui.ViewModels.SettingsViewModel
 import com.example.characterDevelopment.ui.Views.updateTextsLanguage
-import com.example.compose.rally.ui.components.ClickableSimpleTextFieldWithConfirmation
+import com.example.characterDevelopment.ui.components.ClickableSimpleTextFieldWithConfirmation
 
 @Composable
 fun ConfigurationScreen(charactersVm: CharacterCreatorViewModel, settingsVm: SettingsViewModel) {
