+++ app/src/main/java/com/example/characterDevelopment/ui/components/CharacterLevelScreenComponents.kt
@@ -1,4 +1,3 @@
-
 package com.example.characterDevelopment.ui.components
 
 import androidx.compose.foundation.layout.Box
@@ -17,7 +16,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import com.example.compose.rally.ui.components.AnimatedCircle
 
 /**
  * Generic component used by the character level screen. Shows a circle filled depending on the relation between the circleValue and the maxValue. Then add custom rows
@@ -26,19 +24,18 @@ import com.example.compose.rally.ui.components.AnimatedCircle
 fun CharacterInformationBody(
     modifier: Modifier = Modifier,
     circleValue: Int,
-    colorsCircle: List <Color>,
-    circleMaxNumber : Float,
+    colorsCircle: List<Color>,
+    circleMaxNumber: Float,
     circleLabel: String,
     rows: @Composable () -> Unit
 
 ) {
-
-    var value = circleValue / circleMaxNumber
+//calculate the completed % of the circle
+    val value = circleValue / circleMaxNumber
     Column(modifier = modifier.verticalScroll(rememberScrollState())) {
         Box(Modifier.padding(16.dp)) {
             AnimatedCircle(
-
-            listOf(1 - value, value),
+                listOf(1 - value, value),
                 colorsCircle,
                 Modifier
                     .height(300.dp)
@@ -46,12 +43,14 @@ fun CharacterInformationBody(
                     .fillMaxWidth()
             )
             Column(modifier = Modifier.align(Alignment.Center)) {
+//The title of the circle
                 Text(
                     text = circleLabel,
                     style = MaterialTheme.typography.h5,
                     modifier = Modifier.align(Alignment.CenterHorizontally)
 
                 )
+//The numerical value that represents the circle
                 Text(
                     text = circleValue.toString(),
                     style = MaterialTheme.typography.h2,
@@ -62,9 +61,8 @@ fun CharacterInformationBody(
         Spacer(Modifier.height(10.dp))
         Card {
             Column(modifier = Modifier.padding(12.dp)) {
-
+                //All the content you want to show after the circle
                 rows()
-
             }
         }
     }
