+++ app/src/main/java/com/example/characterDevelopment/ui/theme/AppThemes.kt
@@ -4,19 +4,13 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import com.example.characterDevelopment.data.database.entities.Theme
 
-/**
- * A [MaterialTheme] for Rally.
- */
-
-
 @Composable
 fun AppTheme(theme: Theme, content: @Composable () -> Unit) {
-if (theme == Theme.LIGHT) {
-    MaterialTheme(colors = ColorPalette, typography = LightTypography, content = content)
-}
-    else {
-    MaterialTheme(colors = DarkColorPalette, typography = DarkTypography, content = content)
+    if (theme == Theme.LIGHT) {
+        MaterialTheme(colors = ColorPalette, typography = LightTypography, content = content)
+    } else {
+        MaterialTheme(colors = DarkColorPalette, typography = DarkTypography, content = content)
 
-}
+    }
 }
